refactor(cockpit): extract contract gate module
